requirements.txt
flask==2.3.3
twilio==8.10.0
openai==1.2.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.0
